+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@
 - Sửa lỗi uu7 -> uư thay vì ưu
 - Sửa lỗi spell check không thoát dấu móc cho case e-u-w-w
 - Sửa lỗi smartswicth hoạt động không đúng (revert lại cách sử cũ)
+- Sửa lỗi kìaaaaaaaaaaaa
+- Sửa lỗi không hoàn tác dấu sau khi space VNI
+- Sửa lỗi smartswitch và exclude apps không hoạt động đúng trong một số trường hợp
 
 ---
 Cảm ơn bạn đã chọn NexusKey! Mọi góp ý của bạn đều giúp bộ gõ ngày càng hoàn thiện hơn.